Update constant highlighting rules.
We consider something to be a "special constant" if it starts with
"enough" (2 or more in this case) upper-case letters. Any leading
underscores are ignored for the purpose of this definition. Also any
number of underscores and digits are allowed in between the first 2
upper-case letters. To fully satisfy the requirement, the "leading" 2
upper-case letters must be followed by only upper-case letters or
digits until the first underscore.