Bugfix for an apparent infinite loop in grammar.
This is a case of semantically equivalent regex not being
computationally equivalent. The issue also relates to how negative
character classes (e.g. `[^abc]`) are treated. In the end I had to
replace something of the form `(A | B)*` with `A* (BA*)*` to avoid
grinding the regex engine to a halt.

Issue #150.